Upgrade Gradle and AGP, add namespaces to modules
This commit updates the Gradle wrapper to version 8.3 and the Android Gradle Plugin to version 8.1.3.

Additionally, it introduces the `namespace` attribute to the `android` block in the `build.gradle` files for the `app`, `data`, `local`, and `remote` modules. This is now a requirement for AGP.

The `domain` module, being a java-library, does not require a namespace.

IDE configuration files (`.idea/`) have also been updated to reflect these changes, including the Gradle JVM version and bytecode target levels.

The `AndroidManifest.xml` in the `data` module had a trailing slash removed.