test: Update PhotoScreenTest to use photographer name
This commit updates the `photoScreen_displaysPhotos` instrumentation test to align with changes in how photo attribution is displayed.

The test now uses the `photographer` property of the `PhotoData` mock instead of the `original` property. Assertions have been updated to verify that the text "Photo by [photographer name]" is correctly displayed on the screen.

Signed-off-by: Ikechukwu Iroka <martins.iroka@gmail.com>